added authentication

+++ src/service/user/mutation.ts
@@ -1,3 +1,4 @@
+import bcrypt from 'bcryptjs';
 import mongoose from 'mongoose';
 import { sendUserResponse } from "./helper";
 import User from "./model";
@@ -18,23 +19,38 @@ export const UserMutation = {
     }
   },
   updateUser: async (parent: any, args: any, context: any, info: any) => {
+    if(!context.user)  return sendUserResponse(403,"Permission Denied",{})
+    if(context.user.id!==args.updateUserInput.id) return sendUserResponse(403,"Only User Can Update Own Profile",{})
+    
     if(!mongoose.Types.ObjectId.isValid(args.updateUserInput.id)) return sendUserResponse(403,"Invalid ID")
-    else {
-      const password = args.updateUserInput.password
+    else if(args.updateUserInput.password) {
+      const password = args.updateUserInput.password 
       if(password.length<6) return sendUserResponse(403,"Password need to be atleast 6 word")
       else if(password === "password") return sendUserResponse(403,"password can not be password")
+      else {
+        const salt = await bcrypt.genSalt(10)
+        args.updateUserInput.password = await bcrypt.hash(args.updateUserInput.password, salt)
+      }
+    }
+
+    try {
+      const user = await User.findByIdAndUpdate(
+        args.updateUserInput.id,
+        {...args.updateUserInput},
+        { new: true }
+      );
+
+      if (!user) return sendUserResponse(401,"The User With Given ID Not Found",{})
+      return sendUserResponse(200,"Success",user)
+    } catch (error) {
+      console.log(error)
     }
     
-    const user = await User.findByIdAndUpdate(
-      args.updateUserInput.id,
-      {...args.updateUserInput},
-      { new: true }
-    );
-
-    if (!user) return sendUserResponse(401,"The User With Given ID Not Found",{})
-    return sendUserResponse(200,"Success",user)
   },
   deleteUser: async (parent: any, args: any, context: any, info: any) => {
+    if(!context.user)  return sendUserResponse(403,"Permission Denied",{})
+    if(context.user.id!==args.id) return sendUserResponse(403,"Only User Can Delete Own Profile",{})
+
     if(!mongoose.Types.ObjectId.isValid(args.id)) return sendUserResponse(403,"Invalid ID",{})
 
     const user = await User.findByIdAndRemove(args.id);
